+++ src/Test/QuickCheck/Classes.hs
@@ -33,6 +33,7 @@ module Test.QuickCheck.Classes
   , primLaws
   , semigroupLaws
   , commutativeSemigroupLaws
+  , specialSemigroupLaws
 #if HAVE_SEMIRINGS
   , semiringLaws
 #endif
@@ -71,6 +72,12 @@ module Test.QuickCheck.Classes
   , Laws(..)
   , Proxy1(..)
   , Proxy2(..)
+    -- * Special Semigroups
+  , SpecialSemigroup
+  , commutative
+  , exponential
+  , idempotent
+  , rectangularBand
   ) where
 
 --